add examples
