+++ moni/test/node_script.js
@@ -1,5 +1,6 @@
 /* DB setting */
 var DB = require('../../common/db/connection.js');
+var Last = require('../../common/utils/Last.js').Last;
 
 /* casperJs setting */
 var casperProcess = (process.platform === "win32" ? "casperjs.cmd" : "casperjs");
@@ -136,34 +137,3 @@ function callCasper( pages ) {
 }
 
 exports.callCasper = callCasper;
-
-function Last ( initCnt, compareCnt, callback, callbackParam ) {
-	this.count = initCnt || 0;
-	this.compareCount = compareCnt || null;
-	this.callback = callback;
-	this.callbackParam = callbackParam;
-}
-Last.prototype.setCount = function ( count ) {
-	this.count = count;
-}
-Last.prototype.setCompareCount = function ( compareCount ) {
-	this.compareCount = compareCount;
-}
-Last.prototype.setCallback = function ( callback ) {
-	this.callback = callback;
-}
-Last.prototype.setCallbackParam = function ( callbackParam ) {
-	this.callbackParam = callbackParam;
-}
-Last.prototype.trigger = function ( ) {
-	++this.count;
-	console.log( this.count + ' / ' + this.compareCount );
-	if ( this.count == this.compareCount ) {
-		this.action();
-	}
-}
-Last.prototype.action = function ( ) {
-	if ( this.callback && this.callback instanceof Function ) {
-		this.callback( this.callbackParam || null );
-	}
-}
