+++ www/views/ex/select.ejs
@@ -1,0 +1,26 @@
+<!doctype html>
+<html>
+<head>
+</head>
+<body>
+	<h1>db1 select result start  >> </h1>
+	<%=db1  %>
+	<br />
+	<ul>
+	<% for ( var i=0 ; i < db1.length ; i++ ) { %>
+		<li>db1 >>> <%=db1[i].page_no%> / <%=db1[i].msg%></li>
+	<% }%>
+	</ul>
+	<br />
+	<br />
+	<br />
+	<h1>db2 select result start  >> </h1>
+	<%=db2  %>
+	<br />
+	<ul>
+	<% for ( var i=0 ; i < db2.length ; i++ ) { %>
+		<li>db2 >>> <%=db2[i].page_no%> / <%=db2[i].msg%></li>
+	<% }%>
+	</ul>
+</body>
+</html>
